function for checking the token

+++ #justDoIt/task_management/deleteTask.php
@@ -1,6 +1,12 @@
 <?php
     include('../session/session.php');
     include('../database/connection.php');
+    include('../security/checkAuthHash.php');
+
+    if(checkAuthHash($_POST['AuthToken'],$_POST['tokenName'])!=1)
+    {
+        return -1;
+    }
 
     $records = $conn->prepare('DELETE FROM task WHERE id = :id');
     if($records != null)
